requirements.txt
python-dotenv==1.0.1
asyncio==1.6.0
websockets==12.0